requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
paho-mqtt==1.6.1
requests==2.31.0
Pillow>=10.0.0
